changed console prints

+++ src/utils/FileInfo.java
@@ -5,7 +5,7 @@ import app.Node;
 import java.util.List;
 
 // stores info after search necessary to download
-public record FileInfo(String fileName, String fileHash, long fileSize, List<Node.NodeConnectionThread> nodeConnectionThreads) {
+public record FileInfo(String fileName, String fileHash, int fileSize, List<Node.NodeConnectionThread> nodeConnectionThreads) {
 
     @Override
     public String toString() {
